+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Embedded data attachment level changed to "INFO", by @HardNorth
 
 ## [5.2.4]
 ### Changed